% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfam-package.R
\docType{package}
\name{bayesfam-package}
\alias{bayesfam}
\alias{bayesfam-package}
\title{bayesfam: brms Custom Families}
\description{
A collection of custom brms families and some additional functions for existing families originally developed as part of the bayesim package. References: Scholz (2022) \doi{10.5281/zenodo.7099513}; Scholz (2022) \doi{10.48550/ARXIV.2210.06927}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sims1253/bayesfam}
  \item Report bugs at \url{https://github.com/sims1253/bayesfam/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Scholz \email{research.scholz@mailbox.org}

Authors:
\itemize{
  \item Maximilian Scholz \email{research.scholz@mailbox.org}
  \item Paul-Christian Bürkner \email{paul.buerkner@gmail.com}
  \item Yannick Dzubba
}

}
\keyword{internal}
