% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_beta.R
\name{beta_custom}
\alias{beta_custom}
\title{Custom-Beta Stan-implementation in mean parametrization.}
\usage{
beta_custom(link = "logit", link_phi = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_phi}{Link function for phi argument}
}
\value{
BRMS Beta-Custom distribution family
}
\description{
Custom-Beta Stan-implementation in mean parametrization.
}
\examples{
n <- 10000
a <- rnorm(n)
data <- list(a = a, y = bayesim::rbeta_custom(n, bayesim::inv_logit(0.5 * a + 1), 2))
fit1 <- brm(y ~ 1 + a,
  data = data, family = bayesim::beta_custom(),
  stanvars = bayesim::beta_custom()$stanvars, backend = "cmdstan"
)
plot(fit1)
}
