% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{betaprime}
\alias{betaprime}
\title{Title}
\usage{
betaprime(link = "log", link_phi = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_beta}{Link function for beta argument}
}
\value{
BRMS beta-prime distribution family
}
\description{
Title
}
\examples{
data <- list(a = a, y = bayesim::rbetaprime(
  n,
  exp(0.5 * rnorm(1000) + 1), 2
))
fit1 <- brm(y ~ 1 + a,
  data = data, family = bayesim::betaprime(),
  stanvars = bayesim::betaprime()$stanvars, backend = "cmdstan"
)
plot(fit1)
}
