% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchitnormal.R
\name{cauchitnormal}
\alias{cauchitnormal}
\title{Custom BRMS family Cauchitnormal}
\usage{
cauchitnormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Cauchitnormal BRMS model-object
}
\description{
Custom BRMS family Cauchitnormal
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
a <- rnorm(1000)
data <- list(a = a, y = rcauchitnormal(1000, 0.5 * a + 1, 2))
# refresh = 0 supresses chain updates
fit1 <- brms::brm(y ~ 1 + a, data = data,
 family = cauchitnormal(), stanvars = cauchitnormal()$stanvars,
 refresh = 0)
plot(fit1)
}
