% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{custom_gumbel}
\alias{custom_gumbel}
\title{Gumbel brms custom family}
\usage{
custom_gumbel(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}
}
\value{
BRMS Gumbel distribution family
}
\description{
Gumbel brms custom family
}
\examples{
a <- rnorm(n = 1000)
data <- list(a = a, y = rgumbel(n = 1000, mu = a + 2, sigma = 2))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
  family = custom_gumbel(), stanvars = custom_gumbel()$stanvars,
  refresh = 0)
plot(fit)
}
