% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_mean.R
\name{dbeta_mean}
\alias{dbeta_mean}
\title{Mean parameterization the beta pdf.}
\usage{
dbeta_mean(x, mu, phi, log = FALSE)
}
\arguments{
\item{x}{x-value, x e (0, 1)}

\item{mu}{Mean parameter, mu e (0, 1)}

\item{phi}{Precision parameter, phi > 0}

\item{log}{Optional argument. If TRUE, returns log(pdf). Normally False.}
}
\value{
PDF of custom Beta Distribution
}
\description{
Mean parameterization the beta pdf.
}
\details{
The Beta Distribution has Density
\deqn{f(y | \mu, \phi) = \frac{\Gamma(\phi) x^{\mu\phi - 1} (1 - x)^{(1-\mu)\phi}}{\Gamma(\mu\phi)\Gamma((1 - \mu)\phi)} }

With parameterisation of the usual Beta-Distribution's shape parameters a and b as:
\deqn{a := \mu\phi, b := (1 - \mu)\phi}
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 1000)
plot(x, dbeta_mean(x, mu = 0.5, phi = 1), type = "l")
}
