% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchitnormal.R
\name{dcauchitnormal}
\alias{dcauchitnormal}
\title{Cauchitnormal Density distribution.}
\usage{
dcauchitnormal(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Value space of the distribution, x e (0, 1)}

\item{mu}{Median parameter, mu is already Cauchit-transformed, mu unbound}

\item{sigma}{Sigma shape parameter, sigma >= 0}

\item{log}{Bool argument, if true, returns the logarithmic density}
}
\value{
Normal Distribution Density with Cauchit link function
}
\description{
Cauchitnormal Density distribution.
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 1000)
plot(x, dcauchitnormal(x, mu = 0.5, sigma = 2), type = "l")
}
