% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_weibull.R
\name{ddiscrete_weibull}
\alias{ddiscrete_weibull}
\title{Probability density function for the discrete Weibull distribution}
\source{
Scholz, F., & Works, B. P. (1996). Maximum likelihood estimation for
        type I censored Weibull data including covariates. ISSTECH-96-022,
        Boeing Information and Support Services.
}
\usage{
ddiscrete_weibull(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Value, x > 0.}

\item{mu}{Location}

\item{sigma}{Scale, sigma > 0.}

\item{log}{Optional argument. If true, returns the log density.}
}
\value{
Density of the pdf given x, mu and sigma.
}
\description{
Probability density function for the discrete Weibull distribution
}
\details{
The location-scale parameterization has the form
\deqn{f(x) = 1 - exp[-exp[\frac{x - \mu}{\sigma}}

Where \deqn{\mu = log(\alpha)} and \deqn{\sigma = \frac{1}{\beta}}
         of for the scale \deqn{\alpha} and shape \degn{\beta} of the standard
         parameterization:
         \deqn{f(x) = 1 - exp[-(\frac{x + 1}{\alpha})^\beta}
}
\examples{
x <- seq(from = 0.1, to = 20, length.out = 1000)
plot(x, dbetaprime(x, mu = 4, phi = 2), type = "l")
}
