% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mean.R
\name{dgamma_mean}
\alias{dgamma_mean}
\title{Mean parameterization of the gamma pdf.}
\usage{
dgamma_mean(x, mu, a, log = FALSE)
}
\arguments{
\item{x}{Value space, x > 0.}

\item{mu}{Mean parameter of the density, mu > 0.}

\item{a}{Shape parameter, a > 0.}

\item{log}{optional argument. If true, returns logarithmic probability. Default = FALSE}
}
\value{
f(x | mu, k)
}
\description{
Mean parameterization of the gamma pdf.
}
\details{
Define rate constant rho as:
\deqn{\rho(\alpha, \mu) = \frac{\alpha}{\mu}}

The Frechet distribution density is defined as
\deqn{f(y) = \frac{y^{\alpha - 1} exp(-\frac{y}{\rho})} {\rho^\alpha \Gamma(\alpha)} }
}
\examples{
x <- seq(from = 0.01, to = 10, length.out = 1000)
plot(x, dgamma_mean(x, mu = 2, a = 2), type = "l")
}
