% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{dgeneralized_normal}
\alias{dgeneralized_normal}
\title{Probability density function for the generalized_normal distribution}
\usage{
dgeneralized_normal(x, mu, sigma, beta, log = FALSE)
}
\arguments{
\item{x}{Value, unbound}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}

\item{beta}{shape, beta > 0}

\item{log}{Optional argument. If true, returns the log density.}
}
\value{
Density of the pdf given x, mu and sigma
}
\description{
Probability density function for the generalized_normal distribution
}
\details{
The beta prime distribution has density
\deqn{f(y | \mu, \sigma, \beta) = \frac{\beta}{2 \beta \Gamma(1/\beta)}exp(-|z|^\beta)}

Where z is the linear transformation
\deqn{z(y, \mu, \sigma) = \frac{y - \mu}{\sigma}}
}
\examples{
x <- seq(from = -5, to = 10, length.out = 1000)
plot(x, dgeneralized_normal(x, mu = 1, sigma = 1, beta = 0.5), type = "l")
}
