% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{dgompertz}
\alias{dgompertz}
\title{Probability density function for the Gompertz distribution, with Median parametrization.}
\usage{
dgompertz(x, mu, beta, log = FALSE)
}
\arguments{
\item{x}{Value}

\item{mu}{Median parameter}

\item{beta}{Scale parameter}

\item{log}{Optional argument. If TRUE, returns log(pdf).}
}
\value{
f(x | mu, eta)
}
\description{
Probability density function for the Gompertz distribution, with Median parametrization.
}
\details{
PDF of Gompertz implementation, with constant b:
\deqn{b(\mu,\eta) := (1 / \mu) * log1p((-1 / \eta) * log(0.5))}
\deqn{f(x) = \eta*b*exp(\eta + bx - \eta * e^{bx})}
}
\examples{
x <- seq(from = 0.1, to = 5, length.out = 100)
plot(x, dgompertz(x, mu = 2, beta = 4), type = "l")
}
