% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_mean.R
\name{dgumbel_mean}
\alias{dgumbel_mean}
\title{Probability density function for the gumbel distribution}
\usage{
dgumbel_mean(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Value, unbound}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}

\item{log}{Optional argument. If true, returns the log density.}
}
\value{
Density of the pdf given x, mu and sigma
}
\description{
Probability density function for the gumbel distribution
}
\details{
The beta prime distribution has density
\deqn{f(y | \mu, \sigma) = \frac{1}{\sigma} exp(-(z + e^{-z}))}

Where z is the linear transformation
\deqn{z(y, \mu, \sigma) = \frac{y - \mu}{\sigma} + \gamma}

Where $\gamma$ refers to the Euler-Mascheroni constant
}
\examples{
x <- seq(from = -5, to = 10, length.out = 1000)
plot(x, dgumbel_mean(x, mu = 2, sigma = 2), type = "l")
}
