% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{dinverse_burr}
\alias{dinverse_burr}
\title{Probability density function for the Inverse Burr distribution, with
Median parametrization.}
\source{
Parameterization follows the actuar package
(\url{https://search.r-project.org/CRAN/refmans/actuar/html/InverseBurr.html}).
Also known as the Dagum distribution.
}
\usage{
dinverse_burr(x, mu, tau, gamma, log = FALSE)
}
\arguments{
\item{x}{Model space, defined for x > 0}

\item{mu}{Median parameter of pdf, mu > 0}

\item{tau}{Shape1 parameter of pdf, tau > 0 (actuar: shape1)}

\item{gamma}{Shape2 parameter of pdf, gamma > 0 (actuar: shape2)}

\item{log}{Optional log argument, if true, return log(pdf)}
}
\value{
PDF of the Inverse Burr distribution
}
\description{
Probability density function for the Inverse Burr distribution, with
Median parametrization.
}
\details{
Define the scale parameter sigma as
\deqn{\sigma(\mu, \tau, \gamma) := \mu \cdot (2^{1/\tau} - 1)^{1/\gamma}}

\deqn{f(y | \mu, \tau, \gamma) = \frac{\tau \gamma (y/\sigma)^{\gamma \tau}}{y [1 + (y/\sigma)^{\gamma}]^{\tau + 1}}}
}
\examples{
x <- seq(from = 0.1, to = 10, length.out = 100)
plot(x, dinverse_burr(x, mu = 2, tau = 2, gamma = 3), type = "l")
}
