% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_weibull.R
\name{discrete_weibull}
\alias{discrete_weibull}
\title{Discrete Weibull brms custom family}
\usage{
discrete_weibull(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function for the location parameter}

\item{link_sigma}{Link function for the scale parameter}
}
\value{
brms discrete Weibull distribution family
}
\description{
Discrete Weibull brms custom family
}
\examples{
a <- rnorm(n = 1000)
data <- list(a = a, y = rbetaprime(n = 1000, mu = exp(0.5 * a + 1), phi = 2))
fit <- brms::brm(
  formula = y ~ 1 + a, data = data,
  family = betaprime(), stanvars = betaprime()$stanvars,
  refresh = 0
)
plot(fit)
}
