% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumaraswamy.R
\name{dkumaraswamy}
\alias{dkumaraswamy}
\title{Kumaraswamy density function in median parametrisation.}
\usage{
dkumaraswamy(x, mu, p, log = FALSE)
}
\arguments{
\item{x}{value space, x e (0, 1)}

\item{mu}{Median parameter of pdf, mu e (0, 1)}

\item{p}{shape parameter, p > 0}

\item{log}{if true, returns log(pdf). Normally FALSE.}
}
\value{
f(x | mu, p)
}
\description{
Kumaraswamy density function in median parametrisation.
}
\details{
\deqn{q(\mu, p) = -\frac{log(2)}{log(1-\mu^p)}}

\deqn{f(y | \mu, p) = pqx^{p-1}(1-x^p)^{q-1}}
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 1000)
plot(x, dkumaraswamy(x, mu = 0.5, p = 2), type = "l")
}
