% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{dlogistic}
\alias{dlogistic}
\title{Probability density function for the logistic distribution}
\usage{
dlogistic(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Value, unbound}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}

\item{log}{Optional argument. If true, returns the log density.}
}
\value{
Density of the pdf given x, mu and sigma
}
\description{
Probability density function for the logistic distribution
}
\details{
The logistic distribution has density
\deqn{f(y | \mu, \sigma) = \frac{e^{-z}}{\sigma(1 + e^{-z})^2}}

Where z is the linear transformation
\deqn{z(y, \mu, \sigma) = \frac{y - \mu}{\sigma}}
}
\examples{
x <- seq(from = -5, to = 10, length.out = 1000)
plot(x, dlogistic(x, mu = 2, sigma = 1), type = "l")
}
