% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{dlognormal_natural}
\alias{dlognormal_natural}
\title{Title}
\usage{
dlognormal_natural(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Value space, x > 0}

\item{mu}{Mean parameter, mu > 0}

\item{sigma}{Shape parameter, sigma unbound}

\item{log}{boolean value to return log-pdf, default FALSE}
}
\description{
Title
}
