% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{dlomax}
\alias{dlomax}
\title{Probability density function for the Lomax distribution, with Mean parametrization.}
\usage{
dlomax(x, mu, alpha, log = FALSE)
}
\arguments{
\item{x}{Model space, defined for x >= 0}

\item{mu}{Mean parameter of pdf, mu > 0}

\item{alpha}{Alpha parameter of pdf, alpha > 1}

\item{log}{Optional log argument, if true, return log(pdf)}
}
\value{
PDF of Lomax Distribution
}
\description{
Probability density function for the Lomax distribution, with Mean parametrization.
}
\examples{
x <- seq(from = 0, to = 10, length.out = n)
plot(x, bayesim::dlomax(x, mu = 2, alpha = 2), type = "l", ylab = "Density", main = "high starting Lomax(mu=2, eta=2)")
}
