% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_normal.R
\name{dpower_normal}
\alias{dpower_normal}
\title{Probability density function for the power_normal distribution
Implementation after Nelson Wayne
\url{https://archive.org/details/computerprogramp4760nels/page/10/mode/2up}}
\usage{
dpower_normal(x, mu, sigma, beta, log = FALSE)
}
\arguments{
\item{x}{Value, unbound}

\item{mu}{Median, unbound}

\item{sigma}{Scale, sigma > 0}

\item{beta}{Shape, beta > 0}

\item{log}{Optional argument. If true, returns the log density.}
}
\value{
Density of the pdf given x, mu and sigma
}
\description{
Probability density function for the power_normal distribution
Implementation after Nelson Wayne
\url{https://archive.org/details/computerprogramp4760nels/page/10/mode/2up}
}
\details{
The beta prime distribution has density
\deqn{f(y | \mu, \sigma) = \frac{e^{-z}}{\sigma(1 + e^{-z})^2}}

Where z is the linear transformation
\deqn{z(y, \mu, \sigma) = \frac{y - \mu}{\sigma}}
}
\examples{
x <- seq(from = -4, to = 4, length.out = 1000)
plot(x, dpower_normal(x, mu = 0, sigma = 1, beta = 4), type = "l")
}
