% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlognormal.R
\name{dsymlognormal}
\alias{dsymlognormal}
\title{symlognormal density function}
\source{
Based on Hafner, D., Pasukonis, J., Ba, J., & Lillicrap, T. (2023).
        Mastering Diverse Domains through World Models.
        (\url{https://doi.org/10.48550/arXiv.2301.04104})
}
\usage{
dsymlognormal(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Value space of the distribution, x e (0, 1)}

\item{mu}{Median parameter, mu is already symlog-transformed, mu unbound}

\item{sigma}{Sigma shape parameter, sigma >= 0}

\item{log}{Bool argument, if true, returns the logarithmic density}
}
\value{
Normal distribution density with symlog link function
}
\description{
symlognormal density function
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 1000)
plot(x, dsymlognormal(x, mu = 0.5, sigma = 2), type = "l")
}
