% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{erf}
\alias{erf}
\title{Gaussion Error function}
\usage{
erf(x)
}
\arguments{
\item{x}{value to be transformed, x unbound}
}
\value{
erf function of x, result e (0, 1)
}
\description{
Gaussion Error function
}
\examples{
x <- seq(from = -2, to = 2, length.out = 100)
plot(x, erf(x), type = "l")
}
