% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exgauss_custom.R
\name{exgauss2}
\alias{exgauss2}
\title{Exgauss brms custom family, custom implementation vs Stan}
\usage{
exgauss2(link = "identity", link_sigma = "log", link_beta = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}

\item{link_beta}{Link function for beta argument}
}
\value{
BRMS Gumbel distribution family
}
\description{
Exgauss brms custom family, custom implementation vs Stan
}
\examples{
a <- rnorm(n = 1000)
data <- list(a = a, y = brms::rexgaussian(n = 1000, mu = a + 2, sigma = 2, beta = 2))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
  family = exgauss2(), stanvars = exgauss2()$stanvars,
  refresh = 0)
plot(fit)
}
