% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{expect_brms_family}
\alias{expect_brms_family}
\title{BRMS family expect recovery. Tries linear baysian model y ~ 1.
Checking of the arguments done in construct_brms.}
\usage{
expect_brms_family(
  n_data_sampels = 1000,
  intercept,
  ref_intercept = NULL,
  aux_par,
  aux2_par = NA,
  rng_link,
  parameter_link,
  family,
  rng,
  aux_name,
  aux2_name = NULL,
  seed = 1235813,
  data_threshold = NULL,
  thresh = 0.05,
  debug = FALSE
)
}
\arguments{
\item{n_data_sampels}{How many samples per chain. Positive integer scalar. Default = 1000.}

\item{intercept}{Intercept for data generating RNG.}

\item{ref_intercept}{Reference intercept to compare model against. If NULL (default) uses the given intercept.}

\item{aux_par}{Auxiliary parameter of each distribution.}

\item{aux2_par}{2nd Auxiliary parameter (if applicable).}

\item{rng_link}{Link function pointer used for data generation. Mainly for transformed normal distributions.}

\item{parameter_link}{Link function pointer for the latent parameters. Used to transform for comparison with ref_intercept}

\item{family}{BRMS family under test.}

\item{rng}{function pointer of bespoke RNG for the family to be tested.}

\item{aux_name}{BRMS string of aux_par argument name. Single string.}

\item{seed}{Seed argument, so that input data is always the same in each test.
BRMS test does not test RNG and is not guaranteed to fit on all data. Positive Integer scalar, Default = 1235813.
Seed is stored before test and restored after it finished. If wants not to use a seed set to NA.}

\item{data_threshold}{Usually unused. But in rare cases, data too close at the boundary may cause trouble.
If so, set a two entry real vector c(lower, upper). If one of them is NA, the data will not be capped for that boundary.
Default = Null, will be in R terms "invisible" and will not cap any input data.}

\item{thresh}{Acceptable threshold for quantiles of recovered arguments.
Scalar or 2-entry real vector within (0, 1).
Vector is used as is, scalar will be interpreted as c(thresh, 1-thresh).
Default = 0.05}

\item{debug}{Scalar Boolean argument, whether debug info is printed or not. Default = False.}
}
\value{
None
}
\description{
BRMS family expect recovery. Tries linear baysian model y ~ 1.
Checking of the arguments done in construct_brms.
}
\examples{
result <- bayesfam:::expect_brms_family(
  intercept = 5,
  aux_par = 2,
  ref_intercept = 5,
  rng_link = identity,
  parameter_link = log,
  family = betaprime,
  rng = rbetaprime,
  aux_name = "phi"
)
print(result)
}
