% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{generalized_normal}
\alias{generalized_normal}
\title{generalized_normal brms custom family}
\usage{
generalized_normal(link = "identity", link_sigma = "log", link_b = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}

\item{link_beta}{Link function for beta argument}
}
\value{
BRMS generalized_normal distribution family
}
\description{
generalized_normal brms custom family
}
\examples{
a <- rnorm(1000)
data <- list(a = a, y = rgeneralized_normal(n = 1000, mu = 0.5 + 1*a, sigma = 2, beta = 4))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
  family = generalized_normal(), stanvars = generalized_normal()$stanvars,
  refresh = 0)
plot(fit)
}
