% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{gompertz}
\alias{gompertz}
\title{Gompertz Stan-implementation in median parametrization.}
\usage{
gompertz(link = "log", link_b = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_eta}{Link function for eta argument}
}
\value{
BRMS gompertz distribution family
}
\description{
Gompertz Stan-implementation in median parametrization.
}
\examples{
data <- list(a = a, y = bayesim::rgompertz(n, exp(0.5 * a + 1), 0.2))
fit1 <- brm(y ~ 1 + a,
  data = data, family = bayesim::gompertz(),
  stanvars = bayesim::gompertz()$stanvars, backend = "cmdstan"
)
plot(fit1)
}
