% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_cloglog}
\alias{inv_cloglog}
\title{Complementary-Log-Log response function}
\usage{
inv_cloglog(x)
}
\arguments{
\item{x}{value of x to be transformed, x unbound}
}
\value{
inverse-cloglog value of x, result is e (0, 1)
}
\description{
Complementary-Log-Log response function
}
\examples{
x <- seq(from = -3, to = 1, length.out = 100)
plot(x, inv_cloglog(x), type = "l")
}
