% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_logit}
\alias{inv_logit}
\title{Logit response function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{value of x to be transformed, x unbound}
}
\value{
inverse logit value of x, result is e (0, 1)
}
\description{
Logit response function
}
\examples{
x <- seq(from = -5, to = 5, length.out = 100)
plot(x, inv_logit(x), type = "l")
}
