% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_softplus}
\alias{inv_softplus}
\title{Softplus response function}
\usage{
inv_softplus(x)
}
\arguments{
\item{x}{value to be transformed, x is unbound}
}
\value{
inv_softplus of x, result is positive unbound
}
\description{
Softplus response function
}
\examples{
x <- seq(from = 0.1, to = 5, length.out = 100)
plot(x, softplus(x), type = "l")
}
