% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inv_symlog}
\alias{inv_symlog}
\title{Symlog response function}
\source{
Based on Hafner, D., Pasukonis, J., Ba, J., & Lillicrap, T. (2023).
Mastering Diverse Domains through World Models.
(\url{https://doi.org/10.48550/arXiv.2301.04104})
}
\usage{
inv_symlog(x)
}
\arguments{
\item{x}{value to be transformed, x is unbound}
}
\value{
inv_symlog of x, result is unbound
}
\description{
Symlog response function
}
\examples{
inv_symlog(0)
inv_symlog(10)
inv_symlog(-10)
}
