% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{inverse_burr}
\alias{inverse_burr}
\title{Custom Inverse Burr brms-implementation in Median parametrization.}
\source{
Parameterization follows the actuar package
(\url{https://search.r-project.org/CRAN/refmans/actuar/html/InverseBurr.html}).
}
\usage{
inverse_burr(link = "log", link_tau = "log", link_gamma = "log")
}
\arguments{
\item{link}{Link function for mu}

\item{link_tau}{Link function for tau argument}

\item{link_gamma}{Link function for gamma argument}
}
\value{
brms Inverse Burr distribution family
}
\description{
Custom Inverse Burr brms-implementation in Median parametrization.
}
\examples{
a <- rnorm(1000)
data <- list(a = a, y = rinverse_burr(1000, exp(0.5 * a + 1), 2, 3))
fit <- brms::brm(
  formula = y ~ 1 + a, data = data,
  family = inverse_burr(), stanvars = inverse_burr()$stanvars,
  refresh = 0
)
plot(fit)
}
