% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{inverse_burr_scale}
\alias{inverse_burr_scale}
\title{Scale of the Inverse Burr distribution in Median parametrization.}
\usage{
inverse_burr_scale(mu, tau, gamma)
}
\arguments{
\item{mu}{Median parameter, mu > 0}

\item{tau}{Shape1 parameter, tau > 0 (actuar: shape1)}

\item{gamma}{Shape2 parameter, gamma > 0 (actuar: shape2)}
}
\value{
The actuar scale parameter sigma(mu, tau, gamma)
}
\description{
Scale of the Inverse Burr distribution in Median parametrization.
}
