% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{isInt_len}
\alias{isInt_len}
\title{Integer vector check}
\usage{
isInt_len(int, len = 1)
}
\arguments{
\item{int}{Integer vector to be checked}

\item{len}{Length of vector, default argument is 1}
}
\value{
Boolean, whether int was Integer and of correct size
}
\description{
Integer vector check
}
\examples{
bayesfam:::isInt_len(c(1, 2), 2) # should be TRUE
bayesfam:::isInt_len(1, 2) # should be FALSE, wrong length
}
