% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{isLogic_len}
\alias{isLogic_len}
\title{Boolean vector check}
\usage{
isLogic_len(logic, len = 1)
}
\arguments{
\item{logic}{Logic vector to be checked}

\item{len}{Length of vector, default argument is 1}
}
\value{
Boolean, whether logic was Boolean and of correct size
}
\description{
Boolean vector check
}
\examples{
bayesfam:::isLogic_len(c(TRUE, FALSE), 2) # should be TRUE
bayesfam:::isLogic_len(0, len = 1) # should be FALSE, 0 and 1 are numeric
}
