% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{isNum_len}
\alias{isNum_len}
\title{Check if a vector is numeric, has no na entries and has length len}
\usage{
isNum_len(x, len = 1)
}
\arguments{
\item{x}{Numeric vector to be checked}

\item{len}{Length of vector, default argument is 1}
}
\value{
Boolean, whether x was numeric and of correct size
}
\description{
Check if a vector is numeric, has no na entries and has length len
}
\examples{
bayesfam:::isNum_len(c(1.1, 2.2), 2) # should be TRUE
bayesfam:::isNum_len(0.2) # should be TRUE

}
