% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{lenEqual}
\alias{lenEqual}
\title{Length and data check function}
\usage{
lenEqual(
  list_of_vectors,
  scalars_allowed = FALSE,
  type_check = NULL,
  na_allowed = FALSE
)
}
\arguments{
\item{list_of_vectors}{List of vectors of any type you want}

\item{scalars_allowed}{In many applications, mixing scalars with vectors may be fine,
so this argument controls, if scalars (length == 1) would be fine to. Default = FALSE}

\item{type_check}{Function pointer argument of type to be checked. All vectors have to
be of this type, if defined. Default = NULL}

\item{na_allowed}{True if test shall pass with NAs in the data.}
}
\value{
boolean, given the arguments above
}
\description{
Length and data check function
}
\examples{
va <- c(1, 2, 3)
vb <- c(4, 5, 6)
vc <- c(7, 8)
bayesfam:::lenEqual(list(va, vb)) # both got 3 entries
bayesfam:::lenEqual(list(va, vb, vc)) # not all vectors have the same number of entries
}
