% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{limit_data}
\alias{limit_data}
\title{Data limit function}
\usage{
limit_data(data, limits)
}
\arguments{
\item{data}{Data to be limited}

\item{limits}{Limits to be used. Vector with 2 real entries, limits[1] <= limits[2]
If the lower bound does not have to be restricted, set it to NA and vice versa.
Sets data outside those bounds to those bounds.}
}
\value{
data limited by the limits
}
\description{
Data limit function
}
\examples{
input <- c(1, 2, 3, 4)
print(bayesfam:::limit_data(input, c(2, 3)))   # lower and upper bounds
print(bayesfam:::limit_data(input, c(2, NA)))  # only lower bound
}
