% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_beta.R
\name{log_lik_beta}
\alias{log_lik_beta}
\title{Log-Likelihood vignette for the Custom-Beta distribution, with Mean parametrization.}
\usage{
log_lik_beta(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of Beta-Custom given data in prep
}
\description{
Log-Likelihood vignette for the Custom-Beta distribution, with Mean parametrization.
}
