% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{log_lik_betaprime}
\alias{log_lik_betaprime}
\title{Log-Likelihood of the beta prime distribution}
\usage{
log_lik_betaprime(i, prep)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}
}
\value{
Log-Likelihood of beta prime given data in prep
}
\description{
Log-Likelihood of the beta prime distribution
}
