% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchitnormal.R
\name{log_lik_cauchitnormal}
\alias{log_lik_cauchitnormal}
\title{Log-Likelihood vignette for the Cauchitnormal Distribution}
\usage{
log_lik_cauchitnormal(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}
}
\value{
log_likelihood of the Cauchitnormal Distribution, given some brms data.
}
\description{
Log-Likelihood vignette for the Cauchitnormal Distribution
}
