% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglognormal.R
\name{log_lik_cloglognormal}
\alias{log_lik_cloglognormal}
\title{Log-Likelihood vignette for the Chauchitnormal distribution, with Median parametrization.}
\usage{
log_lik_cloglognormal(i, prep)
}
\arguments{
\item{i}{indices}

\item{prep}{brms data}
}
\value{
log_lik
}
\description{
Log-Likelihood vignette for the Chauchitnormal distribution, with Median parametrization.
}
