% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_weibull.R
\name{log_lik_discrete_weibull}
\alias{log_lik_discrete_weibull}
\title{Log-Likelihood of the discrete Weibull distribution}
\usage{
log_lik_discrete_weibull(i, prep)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}
}
\value{
Log-Likelihood of discrete Weibull given data in prep
}
\description{
Log-Likelihood of the discrete Weibull distribution
}
