% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exgauss_custom.R
\name{log_lik_exgauss2}
\alias{log_lik_exgauss2}
\title{Log-Likelihood of the gumbel distribution}
\usage{
log_lik_exgauss2(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of gumbel given data in prep
}
\description{
Log-Likelihood of the gumbel distribution
}
