% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_gamma.R
\name{log_lik_generalized_gamma}
\alias{log_lik_generalized_gamma}
\title{Log-Likelihood of the generalized gamma distribution}
\usage{
log_lik_generalized_gamma(i, prep)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}
}
\value{
Log-Likelihood of generalized gamma given data in prep
}
\description{
Log-Likelihood of the generalized gamma distribution
}
