% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{log_lik_generalized_normal}
\alias{log_lik_generalized_normal}
\title{Log-Likelihood of the generalized_normal distribution}
\usage{
log_lik_generalized_normal(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of generalized_normal given data in prep
}
\description{
Log-Likelihood of the generalized_normal distribution
}
