% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{log_lik_gumbel}
\alias{log_lik_gumbel}
\title{Log-Likelihood of the gumbel distribution}
\usage{
log_lik_gumbel(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of gumbel given data in prep
}
\description{
Log-Likelihood of the gumbel distribution
}
