% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{log_lik_inverse_burr}
\alias{log_lik_inverse_burr}
\title{Log-Likelihood vignette for the Inverse Burr distribution, with Median
parametrization.}
\usage{
log_lik_inverse_burr(i, prep)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}
}
\value{
Log-Likelihood of Inverse Burr given data in prep
}
\description{
Log-Likelihood vignette for the Inverse Burr distribution, with Median
parametrization.
}
