% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{log_lik_logistic}
\alias{log_lik_logistic}
\title{Log-Likelihood of the logistic distribution}
\usage{
log_lik_logistic(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of logistic given data in prep
}
\description{
Log-Likelihood of the logistic distribution
}
