% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnormal.R
\name{log_lik_logitnormal}
\alias{log_lik_logitnormal}
\title{Log-Likelihood vignette for the Logitnormal distribution, in Median parametrization.}
\usage{
log_lik_logitnormal(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}
}
\value{
log_likelihood of the Logitnormal distribution, given some brms data.
}
\description{
Log-Likelihood vignette for the Logitnormal distribution, in Median parametrization.
}
