% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{log_lik_lognormal_natural}
\alias{log_lik_lognormal_natural}
\title{Logarithmic Density BRMS vignette of Lognormal Natural Likelihood}
\usage{
log_lik_lognormal_natural(i, prep)
}
\arguments{
\item{i}{Indices of BRMS data}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of BRMS data
}
\description{
Logarithmic Density BRMS vignette of Lognormal Natural Likelihood
}
