% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_normal.R
\name{log_lik_power_normal}
\alias{log_lik_power_normal}
\title{Log-Likelihood of the power_normal distribution}
\usage{
log_lik_power_normal(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of power_normal given data in prep
}
\description{
Log-Likelihood of the power_normal distribution
}
