% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{log_lik_shifted_inv_gaussian}
\alias{log_lik_shifted_inv_gaussian}
\title{Log-Likelihood of the shifted inverse Gaussian family}
\usage{
log_lik_shifted_inv_gaussian(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}
}
\value{
Log-Likelihood of brms data
}
\description{
Log-Likelihood of the shifted inverse Gaussian family
}
