% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{log_lik_simplex}
\alias{log_lik_simplex}
\title{Log-Likelihood vignette for the Simplex distribution, in Median parametrization.}
\usage{
log_lik_simplex(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of Simplex given data in prep
}
\description{
Log-Likelihood vignette for the Simplex distribution, in Median parametrization.
}
