% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softplusnormal.R
\name{log_lik_softplusnormal}
\alias{log_lik_softplusnormal}
\title{Log-Likelihood vignette for the Softplus distribution, in Median parametrization.}
\usage{
log_lik_softplusnormal(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{BRMS data}
}
\value{
log_likelihood of the Softplus distribution, given some BRMS data.
}
\description{
Log-Likelihood vignette for the Softplus distribution, in Median parametrization.
}
