% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlognormal.R
\name{log_lik_symlognormal}
\alias{log_lik_symlognormal}
\title{Log-Likelihood for the symlognormal distribution,}
\usage{
log_lik_symlognormal(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}
}
\value{
log_likelihood of the symlognormal distribution, given some brms data.
}
\description{
Log-Likelihood for the symlognormal distribution,
}
