% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{log_lik_unitlindley}
\alias{log_lik_unitlindley}
\title{Logarithmic Density BRMS vignette of Unitlindley Likelihood}
\usage{
log_lik_unitlindley(i, prep)
}
\arguments{
\item{i}{Indices of BRMS data}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of BRMS data
}
\description{
Logarithmic Density BRMS vignette of Unitlindley Likelihood
}
