% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{logistic brms custom family}
\usage{
logistic(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}
}
\value{
BRMS logistic distribution family
}
\description{
logistic brms custom family
}
\examples{
a <- rnorm(n = 1000)
data <- list(a = a, y = rlogistic(n = 1000, mu = a + 2, sigma = 2))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
  family = logistic(), stanvars = logistic()$stanvars,
  refresh = 0)
plot(fit)
}
