% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logit}
\alias{logit}
\title{Logit link function}
\usage{
logit(x)
}
\arguments{
\item{x}{value of x to be transformed, x e (0, 1)}
}
\value{
logit value of x, x unbound
}
\description{
Logit link function
}
\examples{
x <- seq(from = 0.1, to = 0.9, length.out = 100)
plot(x, logit(x), type = "l")
}
