% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{lognormal_natural}
\alias{lognormal_natural}
\title{Lognormal Natural BRMS family}
\usage{
lognormal_natural(link = "log", link_sigma = "log")
}
\arguments{
\item{link}{link for the mean argument, default = log}

\item{link_sigma}{link for the shape argument, default = log}
}
\value{
BRMS model object
}
\description{
Lognormal Natural BRMS family
}
\examples{
a <- rnorm(n = 1000)
data <- list(a = a, y = rlognormal_natural(n = 1000, mu = exp(0.5 * a + 1), sigma = exp(2)))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
 family = lognormal_natural(), stanvars = lognormal_natural()$stanvars,
 refresh = 0)
plot(fit)
}
