% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{lomax}
\alias{lomax}
\title{Lomax Stan-implementation in Mean parametrization.}
\usage{
lomax(link = "log", link_alpha = "log1p")
}
\arguments{
\item{link}{Link function for function}

\item{link_alpha}{Link function for eta argument}
}
\value{
BRMS Lomax distribution family
}
\description{
Lomax Stan-implementation in Mean parametrization.
}
\examples{
data <- list(a = a, y = bayesim::rlomax(n, exp(0.5 * a + 1), 2))
fit1 <- brm(y ~ 1 + a,
  data = data, family = bayesim::lomax(),
  stanvars = bayesim::lomax()$stanvars, backend = "cmdstan"
)
plot(fit1)
}
