% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{normale_difference}
\alias{normale_difference}
\title{Uses euler metric for denominator. I like it, that is why. :D}
\usage{
normale_difference(va, vb)
}
\arguments{
\item{va}{Numeric scalar or vector of entries}

\item{vb}{Numeric scalar or vector of entries
If both va and vb are no scalars, their lengths have to be equal}
}
\value{
Vector of normalized differences
}
\description{
Uses euler metric for denominator. I like it, that is why. :D
}
\examples{
bayesfam:::normale_difference(c(1, 1, 1, 1, 1), c(-1, 0, 1, 2, 3))
}
