% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{pinverse_burr}
\alias{pinverse_burr}
\title{Cumulative distribution function for the Inverse Burr distribution, with
Median parametrization.}
\source{
Parameterization follows the actuar package
(\url{https://search.r-project.org/CRAN/refmans/actuar/html/InverseBurr.html}).
}
\usage{
pinverse_burr(q, mu, tau, gamma)
}
\arguments{
\item{q}{Value to evaluate the CDF at, q > 0}

\item{mu}{Median parameter of pdf, mu > 0}

\item{tau}{Shape1 parameter of pdf, tau > 0 (actuar: shape1)}

\item{gamma}{Shape2 parameter of pdf, gamma > 0 (actuar: shape2)}
}
\value{
CDF of the Inverse Burr distribution
}
\description{
Cumulative distribution function for the Inverse Burr distribution, with
Median parametrization.
}
\details{
Define the scale parameter sigma as
\deqn{\sigma(\mu, \tau, \gamma) := \mu \cdot (2^{1/\tau} - 1)^{1/\gamma}}

\deqn{F(y | \mu, \tau, \gamma) = (\frac{(y/\sigma)^{\gamma}}{1 + (y/\sigma)^{\gamma}})^{\tau}}
}
\examples{
x <- seq(from = 0.1, to = 10, length.out = 100)
plot(x, pinverse_burr(x, mu = 2, tau = 2, gamma = 3), type = "l")
}
