% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_beta.R
\name{posterior_epred_beta}
\alias{posterior_epred_beta}
\title{Title}
\usage{
posterior_epred_beta(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Recover the given mean of data prep
}
\description{
Title
}
