% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchitnormal.R
\name{posterior_epred_cauchitnormal}
\alias{posterior_epred_cauchitnormal}
\title{Posterior expected value prediction. Mean undefined for Cauchitnormal.}
\usage{
posterior_epred_cauchitnormal(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Nothing
}
\description{
Posterior expected value prediction. Mean undefined for Cauchitnormal.
}
