% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglognormal.R
\name{posterior_epred_cloglognormal}
\alias{posterior_epred_cloglognormal}
\title{Posterior expected value prediction. Mean undefined for Cloglog-Normal}
\usage{
posterior_epred_cloglognormal(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Nothing
}
\description{
Posterior expected value prediction. Mean undefined for Cloglog-Normal
}
