% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_weibull.R
\name{posterior_epred_discrete_weibull}
\alias{posterior_epred_discrete_weibull}
\title{posterior_epred for the discrete Weibull distribution}
\usage{
posterior_epred_discrete_weibull(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Expected Values of the Posterior Predictive Distribution
}
\description{
posterior_epred for the discrete Weibull distribution
}
