% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exgauss_custom.R
\name{posterior_epred_exgauss2}
\alias{posterior_epred_exgauss2}
\title{posterior_epred for the gumbel distribution}
\usage{
posterior_epred_exgauss2(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Expected Values of the Posterior Predictive Distribution
}
\description{
posterior_epred for the gumbel distribution
}
