% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_gamma.R
\name{posterior_epred_generalized_gamma}
\alias{posterior_epred_generalized_gamma}
\title{posterior_epred for the generalized gamma distribution}
\usage{
posterior_epred_generalized_gamma(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Expected Values of the Posterior Predictive Distribution
}
\description{
posterior_epred for the generalized gamma distribution
}
