% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{posterior_epred_generalized_normal}
\alias{posterior_epred_generalized_normal}
\title{posterior_epred for the generalized_normal distribution}
\usage{
posterior_epred_generalized_normal(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Expected Values of the Posterior Predictive Distribution
}
\description{
posterior_epred for the generalized_normal distribution
}
