% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{posterior_epred_gompertz}
\alias{posterior_epred_gompertz}
\title{Expectation-Predict vignette for the Gompertz distribution, with Median parametrization.
Not defined for the Gompertz family.}
\usage{
posterior_epred_gompertz(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Nothing
}
\description{
Expectation-Predict vignette for the Gompertz distribution, with Median parametrization.
Not defined for the Gompertz family.
}
