% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{posterior_epred_gumbel}
\alias{posterior_epred_gumbel}
\title{posterior_epred for the gumbel distribution}
\usage{
posterior_epred_gumbel(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Expected Values of the Posterior Predictive Distribution
}
\description{
posterior_epred for the gumbel distribution
}
