% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{posterior_epred_inverse_burr}
\alias{posterior_epred_inverse_burr}
\title{Expectation-Predict vignette for the Inverse Burr distribution, with
Median parametrization. The mean only exists for gamma > 1.}
\usage{
posterior_epred_inverse_burr(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Expected value of the Inverse Burr distribution given data in prep
}
\description{
Expectation-Predict vignette for the Inverse Burr distribution, with
Median parametrization. The mean only exists for gamma > 1.
}
