% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_invgaussian.R
\name{posterior_epred_inversegaussian_custom}
\alias{posterior_epred_inversegaussian_custom}
\title{Title}
\usage{
posterior_epred_inversegaussian_custom(prep)
}
\arguments{
\item{prep}{}
}
\description{
Title
}
