% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumaraswamy.R
\name{posterior_epred_kumaraswamy}
\alias{posterior_epred_kumaraswamy}
\title{Posterior expected value prediction of the Kumaraswamy implementation.}
\usage{
posterior_epred_kumaraswamy(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Recover the given mean of data prep
}
\description{
Posterior expected value prediction of the Kumaraswamy implementation.
}
