% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{posterior_epred_logistic}
\alias{posterior_epred_logistic}
\title{posterior_epred for the logistic distribution}
\usage{
posterior_epred_logistic(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Expected Values of the Posterior Predictive Distribution
}
\description{
posterior_epred for the logistic distribution
}
