% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnormal.R
\name{posterior_epred_logitnormal}
\alias{posterior_epred_logitnormal}
\title{Posterior expected value prediction vignette for Logitnormal distribution.}
\usage{
posterior_epred_logitnormal(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Mean of Posterior
}
\description{
Posterior expected value prediction vignette for Logitnormal distribution.
}
