% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{posterior_epred_lognormal_natural}
\alias{posterior_epred_lognormal_natural}
\title{Expectations of posterior predictions of the Lognormal Natural family}
\usage{
posterior_epred_lognormal_natural(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Mean of the posterior predictive distribution
}
\description{
Expectations of posterior predictions of the Lognormal Natural family
}
