% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{posterior_epred_lomax}
\alias{posterior_epred_lomax}
\title{Expectation-Predict vignette for the Lomax distribution, with Mean parametrization.}
\usage{
posterior_epred_lomax(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Recover the given mean of data prep
}
\description{
Expectation-Predict vignette for the Lomax distribution, with Mean parametrization.
}
