% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{posterior_epred_shifted_inv_gaussian}
\alias{posterior_epred_shifted_inv_gaussian}
\title{Expectations of posterior predictions of the Shifted Inverse Gaussian family}
\usage{
posterior_epred_shifted_inv_gaussian(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Mean of the posterior predictive distribution
}
\description{
Expectations of posterior predictions of the Shifted Inverse Gaussian family
}
