% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{posterior_epred_simplex}
\alias{posterior_epred_simplex}
\title{Posterior expected value prediction of the Simplex implementation.}
\usage{
posterior_epred_simplex(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Recover the given mean of data prep
}
\description{
Posterior expected value prediction of the Simplex implementation.
}
