% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softplusnormal.R
\name{posterior_epred_softplusnormal}
\alias{posterior_epred_softplusnormal}
\title{Posterior expected value prediction vignette for Softplus distribution.}
\usage{
posterior_epred_softplusnormal(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Mean of Posterior
}
\description{
Posterior expected value prediction vignette for Softplus distribution.
}
