% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlognormal.R
\name{posterior_epred_symlognormal}
\alias{posterior_epred_symlognormal}
\title{Posterior epred for symlognormal distribution.}
\usage{
posterior_epred_symlognormal(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Mean of Posterior
}
\description{
Posterior epred for symlognormal distribution.
}
