% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{posterior_epred_unitlindley}
\alias{posterior_epred_unitlindley}
\title{Posterior mean BRMS vingette of Unitlindley Likelihood}
\usage{
posterior_epred_unitlindley(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Mean of the Unitlindley posterior
}
\description{
Posterior mean BRMS vingette of Unitlindley Likelihood
}
