% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_beta.R
\name{posterior_predict_beta}
\alias{posterior_predict_beta}
\title{Posterior predictionLog-Likelihood vignette for the Custom-Beta distribution, with Mean parametrization.}
\usage{
posterior_predict_beta(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{}
}
\value{
Posterior prediction of Beta-Custom, given data in prep
}
\description{
Posterior predictionLog-Likelihood vignette for the Custom-Beta distribution, with Mean parametrization.
}
