% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{posterior_predict_betaprime}
\alias{posterior_predict_betaprime}
\title{Title}
\usage{
posterior_predict_betaprime(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{}
}
\value{
Posterior prediction of beta-prime, given data in prep
}
\description{
Title
}
