% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchitnormal.R
\name{posterior_predict_cauchitnormal}
\alias{posterior_predict_cauchitnormal}
\title{Posterior-predict vignette for the Cauchitnormal Distribution}
\usage{
posterior_predict_cauchitnormal(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}

\item{...}{}
}
\value{
The posterior prediction of the Cauchitnormal Distribution, given some brms data.
}
\description{
Posterior-predict vignette for the Cauchitnormal Distribution
}
