% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglognormal.R
\name{posterior_predict_cloglognormal}
\alias{posterior_predict_cloglognormal}
\title{Posterior-predict vignette for the Chauchitnormal distribution, with Median parametrization.}
\usage{
posterior_predict_cloglognormal(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}

\item{...}{}
}
\value{
Posterior prediction of the data
}
\description{
Posterior-predict vignette for the Chauchitnormal distribution, with Median parametrization.
}
