% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_weibull.R
\name{posterior_predict_discrete_weibull}
\alias{posterior_predict_discrete_weibull}
\title{posterior_predict for the discrete Weibull distribution}
\usage{
posterior_predict_discrete_weibull(i, prep, ...)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}

\item{...}{Catchall argument}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
posterior_predict for the discrete Weibull distribution
}
