% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exgauss_custom.R
\name{posterior_predict_exgauss2}
\alias{posterior_predict_exgauss2}
\title{posterior_predict for the gumbel distribution}
\usage{
posterior_predict_exgauss2(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{catchall argument}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
posterior_predict for the gumbel distribution
}
