% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_gamma.R
\name{posterior_predict_generalized_gamma}
\alias{posterior_predict_generalized_gamma}
\title{posterior_predict for the generalized gamma distribution}
\usage{
posterior_predict_generalized_gamma(i, prep, ...)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}

\item{...}{Catchall argument}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
posterior_predict for the generalized gamma distribution
}
