% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{posterior_predict_gompertz}
\alias{posterior_predict_gompertz}
\title{Posterior-Prediction vignette for the Gompertz distribution, with Median parametrization.}
\usage{
posterior_predict_gompertz(i, prep, ...)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}

\item{...}{}
}
\value{
Posterior prediction of gompertz, given data in prep
}
\description{
Posterior-Prediction vignette for the Gompertz distribution, with Median parametrization.
}
