% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{posterior_predict_gumbel}
\alias{posterior_predict_gumbel}
\title{posterior_predict for the gumbel distribution}
\usage{
posterior_predict_gumbel(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
posterior_predict for the gumbel distribution
}
