% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{posterior_predict_inverse_burr}
\alias{posterior_predict_inverse_burr}
\title{Posterior-Prediction vignette for the Inverse Burr distribution, with
Median parametrization.}
\usage{
posterior_predict_inverse_burr(i, prep, ...)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}

\item{...}{Catchall argument}
}
\value{
Posterior prediction of Inverse Burr, given data in prep
}
\description{
Posterior-Prediction vignette for the Inverse Burr distribution, with
Median parametrization.
}
