% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_invgaussian.R
\name{posterior_predict_inversegaussian_custom}
\alias{posterior_predict_inversegaussian_custom}
\title{Title}
\usage{
posterior_predict_inversegaussian_custom(i, prep, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
