% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{posterior_predict_logistic}
\alias{posterior_predict_logistic}
\title{posterior_predict for the logistic distribution}
\usage{
posterior_predict_logistic(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
posterior_predict for the logistic distribution
}
