% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnormal.R
\name{posterior_predict_logitnormal}
\alias{posterior_predict_logitnormal}
\title{Posterior-predict vignette for the Logitnormal distribution, with Median parametrization.}
\usage{
posterior_predict_logitnormal(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{BRMS data}

\item{...}{catchall}
}
\value{
The posterior prediction of the Logitnormal distribution, given some BRMS data.
}
\description{
Posterior-predict vignette for the Logitnormal distribution, with Median parametrization.
}
