% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{posterior_predict_lognormal_natural}
\alias{posterior_predict_lognormal_natural}
\title{Posterior Prediction of the Lognormal Natural family}
\usage{
posterior_predict_lognormal_natural(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}

\item{...}{Catchall argument}
}
\value{
Draws from the posterior predictive distribution
}
\description{
Posterior Prediction of the Lognormal Natural family
}
