% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{posterior_predict_lomax}
\alias{posterior_predict_lomax}
\title{Posterior-Prediction vignette for the Lomax distribution, with Mean parametrization.}
\usage{
posterior_predict_lomax(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{}
}
\value{
Posterior prediction of Lomax, given data in prep
}
\description{
Posterior-Prediction vignette for the Lomax distribution, with Mean parametrization.
}
