% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_normal.R
\name{posterior_predict_power_normal}
\alias{posterior_predict_power_normal}
\title{posterior_predict for the power_normal distribution}
\usage{
posterior_predict_power_normal(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
posterior_predict for the power_normal distribution
}
