% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{posterior_predict_shifted_inv_gaussian}
\alias{posterior_predict_shifted_inv_gaussian}
\title{Posterior Prediction of the Shifted Inverse Gaussian family}
\usage{
posterior_predict_shifted_inv_gaussian(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}

\item{...}{Catchall argument}
}
\value{
Draws from the posterior predictive distribution
}
\description{
Posterior Prediction of the Shifted Inverse Gaussian family
}
