% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlognormal.R
\name{posterior_predict_symlognormal}
\alias{posterior_predict_symlognormal}
\title{Posterior-predict for the symlognormal distribution, with Median parametrization.}
\usage{
posterior_predict_symlognormal(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}

\item{...}{catchall}
}
\value{
The posterior prediction of the symlognormal distribution, given some brms data.
}
\description{
Posterior-predict for the symlognormal distribution, with Median parametrization.
}
