% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{posterior_predict_unitlindley}
\alias{posterior_predict_unitlindley}
\title{Posterior predictions of the Unit Lindley family}
\usage{
posterior_predict_unitlindley(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}

\item{...}{Catchall argument}
}
\value{
Draws from the posterior predictive distribution
}
\description{
Posterior predictions of the Unit Lindley family
}
