% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_normal.R
\name{power_normal}
\alias{power_normal}
\title{power_normal brms custom family}
\usage{
power_normal(link = "identity", link_sigma = "log", link_beta = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}
}
\value{
BRMS power_normal distribution family
}
\description{
power_normal brms custom family
}
\examples{
a <- rnorm(n = 1000)
data <- list(a = a, y = rpower_normal(n = 1000, mu = a + 1, sigma = 2, beta = 10))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
  family = power_normal(), stanvars = power_normal()$stanvars,
  refresh = 0)
plot(fit)
}
