% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_normal.R
\name{power_zf}
\alias{power_zf}
\title{Power transformed qnorm (zf-parameter used by Nelson Wayne)}
\usage{
power_zf(p, beta)
}
\arguments{
\item{p}{percentile to be calculated, 0 < p < 1}

\item{beta}{shape argument, beta > 0}
}
\value{
zf argument used in power-normal computation
}
\description{
Power transformed qnorm (zf-parameter used by Nelson Wayne)
}
