% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_weibull.R
\name{qdiscrete_weibull}
\alias{qdiscrete_weibull}
\title{Quantile function of the discrete Weibull distribution}
\usage{
qdiscrete_weibull(p, mu, sigma)
}
\arguments{
\item{p}{Probabilities, for which to calculate the quantiles}

\item{mu}{Location}

\item{sigma}{Scale, sigma > 0.}
}
\value{
Quantiles of the discrete Weibull distribution
}
\description{
Quantile function of the discrete Weibull distribution
}
\examples{
x <- seq(from = 0, to = 1, length.out = 100)
plot(x, qbetaprime(x, mu = 1, phi = 2), type = "l")
}
