% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_mean.R
\name{qgumbel_mean}
\alias{qgumbel_mean}
\title{Quantile function of the gumbel distribution}
\usage{
qgumbel_mean(p, mu, sigma)
}
\arguments{
\item{p}{quantile value, 0 < p < 1}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}
}
\value{
Quantiles of the gumbel distribution
}
\description{
Quantile function of the gumbel distribution
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 100)
plot(x, qgumbel_mean(x, mu = 2, sigma = 2), type = "l")
}
