% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{qinverse_burr}
\alias{qinverse_burr}
\title{Quantile function for the Inverse Burr distribution, with Median
parametrization.}
\source{
Parameterization follows the actuar package
(\url{https://search.r-project.org/CRAN/refmans/actuar/html/InverseBurr.html}).
}
\usage{
qinverse_burr(p, mu, tau, gamma)
}
\arguments{
\item{p}{Quantile to be calculated, p e (0, 1)}

\item{mu}{Median parameter of pdf, mu > 0}

\item{tau}{Shape1 parameter of pdf, tau > 0 (actuar: shape1)}

\item{gamma}{Shape2 parameter of pdf, gamma > 0 (actuar: shape2)}
}
\value{
Inverse of CDF, calculates a value, given a probability p
}
\description{
Quantile function for the Inverse Burr distribution, with Median
parametrization.
}
\details{
Define the scale parameter sigma as
\deqn{\sigma(\mu, \tau, \gamma) := \mu \cdot (2^{1/\tau} - 1)^{1/\gamma}}

\deqn{Q(p | \mu, \tau, \gamma) = \sigma \cdot (\frac{p^{1/\tau}}{1 - p^{1/\tau}})^{1/\gamma}}
}
\examples{
p <- seq(from = 0.01, to = 0.99, length.out = 100)
plot(p, qinverse_burr(p, mu = 2, tau = 2, gamma = 3), type = "l")
}
