% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{qlogistic}
\alias{qlogistic}
\title{Quantile function of the logistic distribution}
\usage{
qlogistic(p, mu, sigma)
}
\arguments{
\item{p}{quantile value, 0 < p < 1}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}
}
\value{
Quantiles of the logistic distribution
}
\description{
Quantile function of the logistic distribution
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 100)
plot(x, qlogistic(x, mu = 2, sigma = 2), type = "l")
}
