% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{qlomax}
\alias{qlomax}
\title{Quantile function for the Lomax distribution, with Mean parametrization.}
\usage{
qlomax(p, mu, alpha)
}
\arguments{
\item{p}{Quantile to be calculated}

\item{mu}{Median argument of Lomax}

\item{alpha}{Alpha argument of Gompertz}
}
\value{
Inverse of CDF, calculates a value, given a probability p
}
\description{
Quantile function for the Lomax distribution, with Mean parametrization.
}
\examples{
x <- seq(from = 0, to = 1, length.out = 100)
plot(x, bayesim::qgompertz(x, mu = 2, eta = 0.1), type = "l", ylab = "Quantile", main = "apex-after-origin Gompertz(mu=2, eta=0.1)")
}
