% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{qunit_lindley}
\alias{qunit_lindley}
\title{Quantile function of Unit-Lindley Likelihood family}
\usage{
qunit_lindley(p, mu)
}
\arguments{
\item{p}{Percentile of Unit-Lindley family, p e (0, 1)}

\item{mu}{Mean parameter, mu e (0, 1)}
}
\value{
QF(p | mu)
}
\description{
Quantile function of Unit-Lindley Likelihood family
}
\examples{
p <- seq(from=0.1, to=0.9, length.out=100)
plot(p, qunit_lindley(p, mu=0.5))
}
