% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_beta.R
\name{rbeta_custom}
\alias{rbeta_custom}
\title{Custom Beta distribution RNG}
\usage{
rbeta_custom(n, mu, phi)
}
\arguments{
\item{n}{Number of draws.}

\item{mu}{Mean}

\item{phi}{Precision}
}
\value{
n samples beta distributed.
}
\description{
Custom Beta distribution RNG
}
\examples{
samples <- rbeta_custom(1000, 0.5, 1)
hist(samples)
}
