% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_mean.R
\name{rbeta_mean}
\alias{rbeta_mean}
\title{Beta distribution RNG for mean parameterization}
\usage{
rbeta_mean(n, mu, phi)
}
\arguments{
\item{n}{Number of draws.}

\item{mu}{Mean}

\item{phi}{Precision}
}
\value{
n samples Beta distributed.
}
\description{
Beta distribution RNG for mean parameterization
}
\examples{
hist(rbeta_mean(1000, mu = 0.5, phi = 1))
}
