% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{rbetaprime}
\alias{rbetaprime}
\title{Title}
\usage{
rbetaprime(n, mu, phi)
}
\arguments{
\item{n}{Number of beta-prime samples.}

\item{mu}{Mean, mu > 0.}

\item{phi}{Precision, phi > 0.}
}
\value{
Random numbers from the beta-prime distribution.
}
\description{
Title
}
\examples{
y <- bayesim::rbetaprime(100, mu = 1, phi = 2)
}
