% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_weibull.R
\name{rdiscrete_Weibull}
\alias{rdiscrete_Weibull}
\title{RNG for the discrete Weibull distribution}
\usage{
rdiscrete_Weibull(n, mu = 1, sigma = 1)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Location}

\item{sigma}{Scale, sigma > 0.}
}
\value{
Random numbers from the discrete Weibull distribution.
}
\description{
RNG for the discrete Weibull distribution
}
\examples{
hist(rbetaprime(100, mu = 1, phi = 2))
}
