% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exgauss_custom.R
\name{rexgauss_custom}
\alias{rexgauss_custom}
\title{Custom rexgauss with default  values}
\usage{
rexgauss_custom(n, mu = 0, sigma = 1, beta = 1)
}
\arguments{
\item{n}{Number of sampels to draw, has to be a scalar natural}

\item{mu}{Mean argument, mu unbound}

\item{sigma}{Shape parameter, sigma > 0}

\item{beta}{Shape parameter, beta > 0}
}
\value{
Vector of length n in exgaussian distribution
}
\description{
Custom rexgauss with default  values
}
\examples{
hist(rstudent_custom(100, 1, 2, 1))
}
