% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_median.R
\name{rfrechet_median}
\alias{rfrechet_median}
\title{Median parameterization of the Fréchet RNG}
\usage{
rfrechet_median(n, mu = 1, nu = 2)
}
\arguments{
\item{n}{Number samples to draw}

\item{mu}{Median}

\item{nu}{Shape, nu > 0}
}
\value{
n samples in Frechet-Distribution
}
\description{
Median parameterization of the Fréchet RNG
}
\examples{
hist(rfrechet_median(100, mu = 1, nu = 2))
}
