% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mean.R
\name{rgamma_mean}
\alias{rgamma_mean}
\title{Mean parameterization of the gamma RNG.}
\usage{
rgamma_mean(n, mu = 1, a = 1)
}
\arguments{
\item{n}{Number of samples, scalar natural number.}

\item{mu}{Mean parameter, mu > 0.}

\item{a}{Shape parameter, a > 0.}
}
\value{
n Gamma distributed samples.
}
\description{
Mean parameterization of the gamma RNG.
}
\examples{
hist(log(rgamma_mean(10000, mu = 2, a = 1)))
}
