% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{rgeneralized_normal}
\alias{rgeneralized_normal}
\title{RNG for the generalized_normal distribution}
\usage{
rgeneralized_normal(n, mu, sigma, beta)
}
\arguments{
\item{n}{Number of sample}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}

\item{beta}{shape, beta > 0}
}
\value{
Random numbers from the generalized_normal distribution.
}
\description{
RNG for the generalized_normal distribution
}
\examples{
hist(rgeneralized_normal(100, mu = 2, sigma = 2, beta = 2))
}
