% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{rgompertz}
\alias{rgompertz}
\title{RNG function for the Gompertz distribution, with Median parametrization.}
\usage{
rgompertz(n, mu, b)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median argument of Gompertz}

\item{eta}{Eta argument of Gompertz}
}
\value{
A Gompertz distributed RNG vector of size n
}
\description{
RNG function for the Gompertz distribution, with Median parametrization.
}
\examples{
y <- bayesim::rgompertz(n, mu = 2, eta = 0.1)
hist(y, main = c(paste("Median:", median(y)), " for RNG of apex-after-origin Gompertz(mu=2, eta=0.1)"))
}
