% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{rgumbel}
\alias{rgumbel}
\title{RNG for the gumbel distribution}
\usage{
rgumbel(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Mean, mu > 0.}

\item{sigma}{Scale, sigma > 0}
}
\value{
Random numbers from the gumbel distribution.
}
\description{
RNG for the gumbel distribution
}
\examples{
hist(rgumbel(100, mu = 2, sigma = 2))
}
