% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_mean.R
\name{rgumbel_mean}
\alias{rgumbel_mean}
\title{RNG for the gumbel distribution}
\usage{
rgumbel_mean(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}
}
\value{
Random numbers from the gumbel distribution.
}
\description{
RNG for the gumbel distribution
}
\examples{
hist(rgumbel_mean(100, mu = 2, sigma = 2))
}
