% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_burr.R
\name{rinverse_burr}
\alias{rinverse_burr}
\title{RNG function for the Inverse Burr distribution, with Median
parametrization.}
\source{
Parameterization follows the actuar package
(\url{https://search.r-project.org/CRAN/refmans/actuar/html/InverseBurr.html}).
}
\usage{
rinverse_burr(n, mu = 1, tau = 1, gamma = 1)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median parameter, mu > 0}

\item{tau}{Shape1 parameter, tau > 0 (actuar: shape1)}

\item{gamma}{Shape2 parameter, gamma > 0 (actuar: shape2)}
}
\value{
An Inverse Burr distributed RNG vector of size n
}
\description{
RNG function for the Inverse Burr distribution, with Median
parametrization.
}
\examples{
hist(rinverse_burr(1000, mu = 2, tau = 2, gamma = 3))
}
