% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{rlogistic}
\alias{rlogistic}
\title{RNG for the logistic distribution}
\usage{
rlogistic(n, mu = 0, sigma = 1)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}
}
\value{
Random numbers from the logistic distribution.
}
\description{
RNG for the logistic distribution
}
\examples{
hist(rlogistic(100, mu = 2, sigma = 2))
}
