% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{rlomax}
\alias{rlomax}
\title{RNG function for the Lomax distribution, with Mean parametrization.}
\usage{
rlomax(n, mu, alpha)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median argument of Lomax}

\item{alpha}{Eta argument of Lomax}
}
\value{
A Lomax distributed RNG vector of size n
}
\description{
RNG function for the Lomax distribution, with Mean parametrization.
}
\examples{
hist(log(rlomax(1000, mu = 1, alpha = 2)))
}
