% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_normal.R
\name{rpower_normal}
\alias{rpower_normal}
\title{RNG for the power_normal distribution}
\usage{
rpower_normal(n, mu, sigma)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Mean, mu > 0.}

\item{sigma}{Scale, sigma > 0}
}
\value{
Random numbers from the power_normal distribution.
}
\description{
RNG for the power_normal distribution
}
\examples{
hist(rpower_normal(100, mu = 2, sigma = 2))
}
