% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{rshifted_inv_gaussian}
\alias{rshifted_inv_gaussian}
\title{RNG function of the Shifted Inverse Gauss Likelihood}
\usage{
rshifted_inv_gaussian(n, mu = 1, shape = 1, shift = 0)
}
\arguments{
\item{n}{Number samples to draw, natural scalar}

\item{mu}{Mean parameter? of likelihod function, mu > 0}

\item{shape}{Shape parameter of likelihood, shape unbound}

\item{shift}{Shift paramter of likelihood, shift <= 0}
}
\value{
N samples drawn of shifted inverse Gauss
}
\description{
RNG function of the Shifted Inverse Gauss Likelihood
}
\examples{
hist(rshifted_inv_gaussian(100, 1, 1, -1))
}
