% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softplusnormal.R
\name{rsoftplusnormal}
\alias{rsoftplusnormal}
\title{Softplus RNG-function in median parametrization.}
\usage{
rsoftplusnormal(n, mu, sigma)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median parameter, mu unbound, mu already log transformed}

\item{sigma}{Sigma shape parameter, sigma > 0}
}
\value{
n Softplus distributed samples
}
\description{
Softplus RNG-function in median parametrization.
}
\examples{
hist(rsoftplusnormal(100, 1, 2))
}
