% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_mean.R
\name{rstudent_mean}
\alias{rstudent_mean}
\title{Custom rstudent with default df value
mu and df arguments switched to comply to the Bayesfam allowing compatibility
for Bayesim}
\usage{
rstudent_mean(n, mu = 0, df = 1, sigma = 1)
}
\arguments{
\item{n}{Number of sampels to draw, has to be a scalar natural}

\item{mu}{Mean argument, mu unbound}

\item{df}{Degrees of freedom variable}

\item{sigma}{Shape parameter, sigma > 0}
}
\value{
Vector of length n in student distribution
}
\description{
Custom rstudent with default df value
mu and df arguments switched to comply to the Bayesfam allowing compatibility
for Bayesim
}
\examples{
hist(rstudent_mean(100, 1, 2, 1))
}
