% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlognormal.R
\name{rsymlognormal}
\alias{rsymlognormal}
\title{symlognormal RNG-function}
\source{
Based on Hafner, D., Pasukonis, J., Ba, J., & Lillicrap, T. (2023).
Mastering Diverse Domains through World Models.
(\url{https://doi.org/10.48550/arXiv.2301.04104})
}
\usage{
rsymlognormal(n, mu = 0, sigma = 1)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median parameter, mu unbound, mu already symlog transformed}

\item{sigma}{Sigma shape parameter, sigma > 0}
}
\value{
n symlognormal distributed samples
}
\description{
symlognormal RNG-function
}
\examples{
hist(rsymlognormal(100, 0.5, 2))
}
