% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{runit_lindley}
\alias{runit_lindley}
\title{Unit-Lindley RNG function}
\usage{
runit_lindley(n, mu)
}
\arguments{
\item{n}{Number of draws, scalar natural number}

\item{mu}{Mean argument, mu e (0, 1)}
}
\value{
N number of samples drawn from Unit-Lindley likelihood
}
\description{
Unit-Lindley RNG function
}
\examples{
hist(runit_lindley(100, 0.5))
}
