% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_median.R
\name{rweibull_median}
\alias{rweibull_median}
\title{Median parameterization of the Weibull RNG.}
\usage{
rweibull_median(n, mu = 1, k = 1)
}
\arguments{
\item{n}{Number of samples, scalar natural number.}

\item{mu}{Median parameter, mu > 0.}

\item{k}{Shape parameter, k > 0.}
}
\value{
n Weibull distributed samples.
}
\description{
Median parameterization of the Weibull RNG.
}
\examples{
hist(log(rweibull_mean(10000, mu = 2, k = 1)))
}
