% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_lognormal_uniform.R
\name{shifted_lognormal_uniform_distribution}
\alias{shifted_lognormal_uniform_distribution}
\alias{rshifted_lognormal_uniform}
\alias{dshifted_lognormal_uniform}
\title{Probability density and RNG function for the mixture of shifted lognormal and
uniform distributions.}
\source{
Some background, discussion and examples at
\url{http://www.martinmodrak.cz/2021/04/01/using-brms-to-model-reaction-times-contaminated-with-errors/}
}
\usage{
rshifted_lognormal_uniform(
  n,
  meanlog = 0,
  sdlog = 1,
  mix = 0.1,
  shift = 0,
  max_uniform = 100
)

dshifted_lognormal_uniform(
  y,
  meanlog = 0,
  sdlog = 1,
  mix = 0.1,
  shift = 0,
  max_uniform = 100
)
}
\arguments{
\item{n}{the number of values to draw from the RNG}

\item{meanlog}{the mean of the lognormal component}

\item{sdlog}{the sd of the lognormal component}

\item{mix}{the probability of the value comming from the uniform component}

\item{shift}{the shift the lognormal distribution}

\item{max_uniform}{the maximum value of the uniform component}

\item{y}{the observed value}
}
\description{
Probability density and RNG function for the mixture of shifted lognormal and
uniform distributions.
}
\details{
The  mixture of shifted lognormal and uniform can be described as
\deqn{y_i =
  \begin{cases}
  u_i  & \mathrm{if} \quad z_i = 0 \\
  s_i + r_i  &  \mathrm{if} \quad z_i = 1
  \end{cases}
  \\
  u_i \sim Uniform(0, \alpha) \\
  \log(r_i) \sim Normal(\mu_i, \sigma) \\
  P(z_i = 0) = \theta}

Where θ corresponds to  \code{mix}, α to  \code{max_uniform}
and \eqn{s_i} to  \code{shift}.
}
