% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex}
\alias{simplex}
\title{Simplex BRMS-implementation in median parametrization.}
\usage{
simplex(link = "logit", link_sigma = "identity")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}
}
\value{
BRMS Beta-Custom distribution family
}
\description{
Simplex BRMS-implementation in median parametrization.
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
a <- rnorm(1000)
data <- list(a = a, y = rsimplex(1000, brms::inv_logit_scaled(0.5 * a + 1), 2))
# BBmisc::surpressAll necassary to keep test output clean
BBmisc::suppressAll({
  fit1 <- brms::brm(y ~ 1 + a,
    data = data, family = simplex(),
    stanvars = simplex()$stanvars, backend = "cmdstanr", cores = 4
  )
})
plot(fit1)
}
