% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{softplus}
\alias{softplus}
\title{Softplus link function}
\usage{
softplus(x)
}
\arguments{
\item{x}{value to be transformed, x positive unbound}
}
\value{
softplus function of x, result is unbound
}
\description{
Softplus link function
}
\examples{
x <- seq(from = 0, to = 5, length.out = 100)
plot(x, softplus(x), type = "l")
}
