% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softplusnormal.R
\name{softplusnormal}
\alias{softplusnormal}
\title{Custom brms family Softplus in median parametrization.}
\usage{
softplusnormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Softplus brms model-object
}
\description{
Custom brms family Softplus in median parametrization.
}
\examples{
a <- rnorm(1000)
data <- list(a = a, y = rsoftplusnormal(1000, 0.5 * a + 1, 2))
fit <- brms::brm(
  formula = y ~ 1 + a, data = data,
  family = softplusnormal(), stanvars = softplusnormal()$stanvars,
  refresh = 0
)
plot(fit)
}
