% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{symlog}
\alias{symlog}
\title{Symlog link function}
\source{
Based on Hafner, D., Pasukonis, J., Ba, J., & Lillicrap, T. (2023).
        Mastering Diverse Domains through World Models.
        (\url{https://doi.org/10.48550/arXiv.2301.04104})
}
\usage{
symlog(x)
}
\arguments{
\item{x}{value to be transformed, x is unbound}
}
\value{
symlog of x, result is unbound
}
\description{
Symlog link function
}
\examples{
symlog(0)
symlog(1e10)
symlog(-1e10)
}
