% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlognormal.R
\name{symlognormal}
\alias{symlognormal}
\title{Custom brms family symlog-Normal}
\source{
Based on Hafner, D., Pasukonis, J., Ba, J., & Lillicrap, T. (2023).
        Mastering Diverse Domains through World Models.
        (\url{https://doi.org/10.48550/arXiv.2301.04104})
}
\usage{
symlognormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
symlognormal brms model-object
}
\description{
Custom brms family symlog-Normal
}
\examples{
a <- rnorm(1000)
data <- list(a = a, y = rsymlognormal(1000, 0.5 * a + 1, 2))
fit <- brms::brm(
  formula = y ~ 1 + a, data = data,
  family = symlognormal(), stanvars = symlognormal()$stanvars,
  refresh = 0
)
plot(fit)
}
