% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{test_rng}
\alias{test_rng}
\title{Tests if an RNG can achieve a good enough location often enough}
\usage{
test_rng(
  rng_fun,
  metric_mu,
  n,
  mu_list,
  aux_list,
  mu_eps,
  p_acceptable_failures,
  mu_link = identity,
  relative = FALSE,
  debug = TRUE
)
}
\arguments{
\item{rng_fun}{RNG function under test}

\item{metric_mu}{Metric to be used on RNG data (usually mean or median)}

\item{n}{Sample size for the rng test.}

\item{mu_list}{Metric data used as RNG argument and to be compared to (usually mean or median)}

\item{aux_list}{Auxiliary parameter}

\item{mu_eps}{Acceptable difference of |mu - metric_mu(rng_fun)}

\item{p_acceptable_failures}{Acceptable rate of failure, relative value of difference bigger mu_eps}

\item{mu_link}{Default=identity, optional link-function argument, for example
useful in link-normal-distributions}

\item{relative}{True if the error should be relative to the mu_list, Default = FALSE}

\item{debug}{bool argument, if set will printout the difference, in case of failure. Default = FALSE}
}
\value{
Nothing actually, just wraps the test
}
\description{
Tests if an RNG can achieve a good enough location often enough
}
\examples{
library(testthat)
eps <- 1e-6
mu_list <- seq(from = 1 + eps, to = 20, length.out = 10)
phis <- seq(from = 2 + eps, to = 20, length.out = 10)
bayesfam:::test_rng(
  rng_fun = rbetaprime, metric_mu = mean, n = 10000, mu_list = mu_list,
  aux_list = phis, mu_eps = 0.2, p_acceptable_failures = 0.05
)
}
