% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{aux_limits_lookup}
\alias{aux_limits_lookup}
\title{Lookup for limits of family auxiliary parameters.}
\usage{
aux_limits_lookup(family)
}
\arguments{
\item{family}{The identifier string of a family.}
}
\value{
List containing lower and upper bounds for the auxiliary parameter.
}
\description{
Lookup for limits of family auxiliary parameters.
}
\examples{
aux_limits_lookup("beta")
}
