% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{betaprime}
\alias{betaprime}
\title{Beta prime brms custom family}
\usage{
betaprime(link = "log", link_phi = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_phi}{Link function for beta argument}
}
\value{
brms beta prime distribution family
}
\description{
Beta prime brms custom family
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
a <- rnorm(n = 1000)
data <- list(a = a, y = rbetaprime(n = 1000, mu = exp(0.5 * a + 1), phi = 2))
# BBmisc::surpressAll necassary to keep the test output clean
fit1 <- brms::brm(y ~ 1 + a,
 data = data, family = betaprime(), stanvars = betaprime()$stanvars,
 silent = 2, refresh = 0
)
print(fit1)
plot(fit1)
}
