% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglognormal.R
\name{cloglognormal}
\alias{cloglognormal}
\title{Custom BRMS family Cloglog-Normal in median parametrization.}
\usage{
cloglognormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Cloglog BRMS model-object
}
\description{
Custom BRMS family Cloglog-Normal in median parametrization.
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
raw_data <- rcloglognormal(1000, 0.5, 2)
# cloglognormal does not like values to close to the boundary
cloglog_data <- limit_data(raw_data, c(1e-12, 1 - 1e-12))
# refresh = 0 supresses chain updates
fit1 <- brms::brm(y ~ 1, data = list(y = cloglog_data),
 family = cloglognormal(), stanvars = cloglognormal()$stanvars,
 refresh = 0)
plot(fit1)
}
