% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_median.R
\name{dfrechet_median}
\alias{dfrechet_median}
\title{Median parameterization of the Fréchet pdf.}
\usage{
dfrechet_median(x, mu, nu)
}
\arguments{
\item{x}{x value space, x > 0}

\item{mu}{Median}

\item{nu}{Shape}
}
\value{
dfrechet(x | mu, nu)
}
\description{
Median parameterization of the Fréchet pdf.
}
\details{
Define scale parameter sigma as
\deqn{\sigma(\mu, \nu) := \mu / \Gamma(1 - 1 / \nu)}

The Frechet distribution has density
\deqn{f(y) = (\nu /\sigma) * (y / \sigma)^{-(1 - \nu)} * exp(-(y / \sigma)^{-\nu}) }
}
\examples{
x <- seq(from = 0.1, to = 20, length.out = 1000)
plot(x, dfrechet_median(x, mu = 6, nu = 4), type = "l")
}
