% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{dlognormal_natural}
\alias{dlognormal_natural}
\title{Lognormal Natural density function with value space parameterisation (mu same as x)}
\usage{
dlognormal_natural(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{vector of values, x > 0}

\item{mu}{Mean, mu > 0}

\item{sigma}{Shape, sigma unbound}

\item{log}{logical; if TRUE, log(pdf) is returned}
}
\value{
Density of the Lognormal Natural given x, mu and sigma.
}
\description{
Lognormal Natural density function with value space parameterisation (mu same as x)
}
\examples{
x <- seq(from = 0.1, to = 5, length.out = 100)
plot(x, dlognormal_natural(x, 1, 2))
}
