% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{dlognormal_natural}
\alias{dlognormal_natural}
\title{Lognormal Natural Density function with value space parameterisation (mu same as x)}
\usage{
dlognormal_natural(x, mu, sigma, log = FALSE)
}
\arguments{
\item{x}{Value space, x > 0}

\item{mu}{Mean parameter, mu > 0}

\item{sigma}{Shape parameter, sigma unbound}

\item{log}{boolean value to return log-pdf, default FALSE}
}
\value{
Density of the Lognormal Natural given x, mu and sigma.
}
\description{
Lognormal Natural Density function with value space parameterisation (mu same as x)
}
\examples{
x <- seq(from=0.1, to=5, length.out=100)
plot(x, dlognormal_natural(x, 1, 2))
}
