% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{dlomax}
\alias{dlomax}
\title{Probability density function for the Lomax distribution, with Mean parametrization.}
\usage{
dlomax(x, mu, alpha, log = FALSE)
}
\arguments{
\item{x}{Model space, defined for x >= 0}

\item{mu}{Mean parameter of pdf, mu > 0}

\item{alpha}{Alpha parameter of pdf, alpha > 1}

\item{log}{Optional log argument, if true, return log(pdf)}
}
\value{
PDF of Lomax Distribution
}
\description{
Probability density function for the Lomax distribution, with Mean parametrization.
}
\details{
\deqn{f(y | \mu, \alpha) = \frac{\alpha}{\mu(\alpha-1)} (1 + \frac{y}{\mu(\alpha-1)})^{-\alpha-1}}
}
\examples{
x <- seq(from = 0, to = 10, length.out = 100)
plot(x, dlomax(x, mu = 1, alpha = 2), type = "l")
}
