% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{dshifted_inv_gaussian}
\alias{dshifted_inv_gaussian}
\title{Shifted inverse Gaussian density function}
\usage{
dshifted_inv_gaussian(x, mu, shape, shift, log = FALSE)
}
\arguments{
\item{x}{Value space of likelihood, x > 0}

\item{mu}{Mean , mu > 0}

\item{shape}{Shape, shape unbound}

\item{shift}{Shift, shift >= 0}

\item{log}{logical; if TRUE, log(pdf) is returned}
}
\value{
f(x | mu, shape, shift)
}
\description{
Shifted inverse Gaussian density function
}
\examples{
x <- seq(from = 0.1, to = 10, length.out = 1000)
plot(x, dshifted_inv_gaussian(x, 1, 1, 1))
}
