% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{dunit_lindley}
\alias{dunit_lindley}
\title{Density of the Unit Lindley likelihood}
\usage{
dunit_lindley(x, mu, log = FALSE)
}
\arguments{
\item{x}{vector of values, x e (0, 1)}

\item{mu}{Mean, mu e (0, 1)}

\item{log}{logical; if TRUE, log(pdf) is returned}
}
\value{
f(x | mu)
}
\description{
Density of the Unit Lindley likelihood
}
\examples{
x <- seq(from = 0.1, to = 0.9, length.out = 1000)
plot(x, dunit_lindley(x, 0.5))
}
