% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{generalized_normal}
\alias{generalized_normal}
\title{generalized_normal brms custom family}
\usage{
generalized_normal(link = "identity", link_sigma = "log", link_b = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}
}
\value{
BRMS generalized_normal distribution family
}
\description{
generalized_normal brms custom family
}
\examples{
data <- list(y = rgeneralized_normal(n = 1000, mu = 2, sigma = 2, beta = 4))
fit <- brms::brm(formula = y ~ 1, data = data,
  family = generalized_normal(), stanvars = generalized_normal()$stanvars,
  cores = 4, init = 0.1)
plot(fit)
}
