% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{generalized_normal}
\alias{generalized_normal}
\title{Generalized Normal BRMS family}
\usage{
generalized_normal(link = "identity", link_sigma = "log", link_b = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}

\item{link_b}{Link function for beta argument}
}
\value{
BRMS generalized_normal distribution family
}
\description{
Generalized Normal BRMS family
}
\examples{
data <- list(y = rgeneralized_normal(n = 1000, mu = 2, sigma = 2, beta = 4))
fit <- brms::brm(formula = y ~ 1, data = data,
  family = generalized_normal(), stanvars = generalized_normal()$stanvars,
  init = 0.1)
plot(fit)
}
