% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{gompertz}
\alias{gompertz}
\title{Custom Gompertz BRMS-implementation in median parametrization.}
\usage{
gompertz(link = "log", link_b = "log")
}
\arguments{
\item{link}{Link function for function}

\item{link_b}{Link function for eta argument}
}
\value{
BRMS gompertz distribution family
}
\description{
Custom Gompertz BRMS-implementation in median parametrization.
}
\examples{
a <- rnorm(1000)
data <- list(a = a, y = rgompertz(1000, mu = exp(0.5 * a + 1), beta = 0.1))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
 family = gompertz(), stanvars = gompertz()$stanvars,
 refresh = 0)
plot(fit)
}
