% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{link_lookup}
\alias{link_lookup}
\title{Lookup function for link and repsonse functions via string identifier.}
\usage{
link_lookup(link, family = NULL, inv = FALSE)
}
\arguments{
\item{link}{String identifier for the link function of interest.}

\item{family}{If a transformed normal family is passed, returns the
respective link instead of \code{link}}

\item{inv}{True to return the response function instead of the link function.}
}
\value{
The respective link function.
}
\description{
If a transformed normal likelihood is passed, the respective built-in link will
be returned instead of the identity link that would commonly be used with
transformed normal likelihood families.
}
\examples{

link_lookup("log", "gaussian", FALSE)

link_lookup("identiy", "logitnormal", FALSE)
}
