% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{log_lik_betaprime}
\alias{log_lik_betaprime}
\title{Title}
\usage{
log_lik_betaprime(i, prep)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of betaprime given data in prep
}
\description{
Title
}
