% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{log_lik_lognormal_natural}
\alias{log_lik_lognormal_natural}
\title{Log-Likelihood of the Lognormal Natural family}
\usage{
log_lik_lognormal_natural(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}
}
\value{
Log-Likelihood of brms data
}
\description{
Log-Likelihood of the Lognormal Natural family
}
