% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{log_lik_shifted_inv_gaussian}
\alias{log_lik_shifted_inv_gaussian}
\title{Logarithmic Density BRMS vignette of shifted inverse Gauss Likelihood}
\usage{
log_lik_shifted_inv_gaussian(i, prep)
}
\arguments{
\item{i}{Indices of BRMS data}

\item{prep}{BRMS data}
}
\value{
Log-Likelihood of BRMS data
}
\description{
Logarithmic Density BRMS vignette of shifted inverse Gauss Likelihood
}
