% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{log_lik_unitlindley}
\alias{log_lik_unitlindley}
\title{Log-likelihood of the Unit Lindley family}
\usage{
log_lik_unitlindley(i, prep)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}
}
\value{
Log-Likelihood of brms data
}
\description{
Log-likelihood of the Unit Lindley family
}
