% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnormal.R
\name{logitnormal}
\alias{logitnormal}
\title{Custom BRMS family Logit-Normal in median parametrization.}
\usage{
logitnormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Logitnormal BRMS model-object
}
\description{
Custom BRMS family Logit-Normal in median parametrization.
}
\examples{
a <- rnorm(1000)
data <- list(a = a, y = rlogitnormal(1000, 0.5 * a + 1, 2))
fit <- brms::brm(formula = y ~ 1 + a, data = data,
 family = logitnormal(), stanvars = logitnormal()$stanvars,
 refresh = 0)
plot(fit)
}
