% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logsumexp}
\alias{logsumexp}
\title{Logarithm of the sum of exponentials.}
\source{
https://en.wikipedia.org/wiki/LogSumExp#log-sum-exp_trick_for_log-domain_calculations
}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
log(sum(exp(x)))
}
\description{
A more numerically stable equivalent to \code{log(sum(exp(x)))}
}
