% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{lomax}
\alias{lomax}
\title{Lomax Stan-implementation in Mean parametrization.}
\usage{
lomax(link = "log", link_alpha = "log1p")
}
\arguments{
\item{link}{Link function for function}

\item{link_alpha}{Link function for eta argument}
}
\value{
BRMS Lomax distribution family
}
\description{
Lomax Stan-implementation in Mean parametrization.
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
a <- rnorm(1000)
data <- list(a = a, y = rlomax(1000, exp(0.5 * a + 1), 2))
# BBmisc::surpressAll necassary to keep the test output clean
BBmisc::suppressAll({
  fit1 <- brms::brm(y ~ 1 + a,
    data = data, family = lomax(),
    stanvars = lomax()$stanvars
  )
})
plot(fit1)
}
