% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{posterior_epred_betaprime}
\alias{posterior_epred_betaprime}
\title{posterior_epred for the beta prime distribution}
\usage{
posterior_epred_betaprime(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Expected Values of the Posterior Predictive Distribution
}
\description{
posterior_epred for the beta prime distribution
}
