% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{posterior_epred_lognormal_natural}
\alias{posterior_epred_lognormal_natural}
\title{Posterior mean BRMS vingette of Lognormal Natural Likelihood}
\usage{
posterior_epred_lognormal_natural(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Mean of the Lognormal natural posterior
}
\description{
Posterior mean BRMS vingette of Lognormal Natural Likelihood
}
