% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{posterior_epred_shifted_inv_gaussian}
\alias{posterior_epred_shifted_inv_gaussian}
\title{Posterior mean BRMS vingette of Shifted Inverse Gauss Likelihood}
\usage{
posterior_epred_shifted_inv_gaussian(prep)
}
\arguments{
\item{prep}{BRMS data}
}
\value{
Mean of the shifted inverse Gauss posterior
}
\description{
Posterior mean BRMS vingette of Shifted Inverse Gauss Likelihood
}
