% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{posterior_epred_unitlindley}
\alias{posterior_epred_unitlindley}
\title{Expectations of posterior predictions of the Unit Lindley family}
\usage{
posterior_epred_unitlindley(prep)
}
\arguments{
\item{prep}{brms data}
}
\value{
Mean of the posterior predictive distribution
}
\description{
Expectations of posterior predictions of the Unit Lindley family
}
