% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{posterior_predict_betaprime}
\alias{posterior_predict_betaprime}
\title{posterior_predict for the beta prime distribution}
\usage{
posterior_predict_betaprime(i, prep, ...)
}
\arguments{
\item{i}{BRMS indices}

\item{prep}{BRMS data}

\item{...}{}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
posterior_predict for the beta prime distribution
}
