% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumaraswamy.R
\name{posterior_predict_kumaraswamy}
\alias{posterior_predict_kumaraswamy}
\title{Posterior prediction vignette for the Kumaraswamy distribution, in Median parametrization.}
\usage{
posterior_predict_kumaraswamy(i, prep, ...)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}

\item{...}{}
}
\value{
Posterior prediction of Kumaraswamy, given data in prep
}
\description{
Posterior prediction vignette for the Kumaraswamy distribution, in Median parametrization.
}
