% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_natural.R
\name{posterior_predict_lognormal_natural}
\alias{posterior_predict_lognormal_natural}
\title{Posterior Prediction BRMS Vignette of Lognormal Natural Likelihood}
\usage{
posterior_predict_lognormal_natural(i, prep, ...)
}
\arguments{
\item{i}{Indices of BRMS data}

\item{prep}{BRMS data}

\item{...}{Catchall argument}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
Posterior Prediction BRMS Vignette of Lognormal Natural Likelihood
}
