% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{posterior_predict_shifted_inv_gaussian}
\alias{posterior_predict_shifted_inv_gaussian}
\title{Posterior Prediction BRMS Vignette of Shifted Inverse Gauss}
\usage{
posterior_predict_shifted_inv_gaussian(i, prep, ...)
}
\arguments{
\item{i}{Indices of BRMS data}

\item{prep}{BRMS data}

\item{...}{Catchall argument}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
Posterior Prediction BRMS Vignette of Shifted Inverse Gauss
}
