% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{posterior_predict_simplex}
\alias{posterior_predict_simplex}
\title{Posterior prediction vignette for the Simplex distribution, in Median parametrization.}
\usage{
posterior_predict_simplex(i, prep, ...)
}
\arguments{
\item{i}{brms indices}

\item{prep}{brms data}

\item{...}{}
}
\value{
Posterior prediction of Simplex, given data in prep
}
\description{
Posterior prediction vignette for the Simplex distribution, in Median parametrization.
}
