% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softplusnormal.R
\name{posterior_predict_softplusnormal}
\alias{posterior_predict_softplusnormal}
\title{Posterior-predict vignette for the Softplus distribution, with Median parametrization.}
\usage{
posterior_predict_softplusnormal(i, prep, ...)
}
\arguments{
\item{i}{Indices}

\item{prep}{brms data}

\item{...}{}
}
\value{
The posterior prediction of the Softplus distribution, given some brms data.
}
\description{
Posterior-predict vignette for the Softplus distribution, with Median parametrization.
}
