% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{posterior_predict_unitlindley}
\alias{posterior_predict_unitlindley}
\title{Posterior Prediction BRMS Vignette of Unitlindley Distribution}
\usage{
posterior_predict_unitlindley(i, prep, ...)
}
\arguments{
\item{i}{Indices of BRMS data}

\item{prep}{BRMS data}

\item{...}{Catchall argument}
}
\value{
Draws from the Posterior Predictive Distribution
}
\description{
Posterior Prediction BRMS Vignette of Unitlindley Distribution
}
