% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_normal.R
\name{qgeneralized_normal}
\alias{qgeneralized_normal}
\title{Quantile function of the generalized_normal distribution
\url{https://cran.r-project.org/web/packages/gnorm/vignettes/gnormUse.html}
BTW, don't believe the * found on Wikipedia.
TODO: well, write this here a bit more nicely^^}
\usage{
qgeneralized_normal(p, mu, sigma, beta)
}
\arguments{
\item{p}{quantile value, 0 < p < 1}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}

\item{beta}{shape, beta > 0}
}
\value{
Quantiles of the generalized_normal distribution
}
\description{
Quantile function of the generalized_normal distribution
\url{https://cran.r-project.org/web/packages/gnorm/vignettes/gnormUse.html}
BTW, don't believe the * found on Wikipedia.
TODO: well, write this here a bit more nicely^^
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 100)
plot(x, qgeneralized_normal(x, mu = 2, sigma = 2, beta = 1), type = "l")
}
