% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{qgompertz}
\alias{qgompertz}
\title{Quantile function for the Gompertz distribution, with Median parametrization.}
\usage{
qgompertz(p, mu, beta)
}
\arguments{
\item{p}{Quantile to be calculated}

\item{mu}{Median parameter}

\item{beta}{Scale parameter}
}
\value{
Inverse of CDF, calculates a value, given a probability p
}
\description{
Quantile function for the Gompertz distribution, with Median parametrization.
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 100)
plot(x, qgompertz(x, mu = 10, beta = 1), type = "l")
}
