% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_normal.R
\name{qpower_normal}
\alias{qpower_normal}
\title{Quantile function of the power_normal distribution}
\usage{
qpower_normal(p, mu, sigma)
}
\arguments{
\item{p}{quantile value, 0 < p < 1}

\item{mu}{Mean, unbound}

\item{sigma}{Scale, sigma > 0}
}
\value{
Quantiles of the power_normal distribution
}
\description{
Quantile function of the power_normal distribution
}
\examples{
x <- seq(from = 0.01, to = 0.99, length.out = 100)
plot(x, qpower_normal(x, mu = 2, sigma = 2), type = "l")
}
