% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.R
\name{rbetaprime}
\alias{rbetaprime}
\title{RNG for the beta prime distribution}
\usage{
rbetaprime(n, mu = 1, phi = 1)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Mean, mu > 0.}

\item{phi}{Precision, phi > 0.}
}
\value{
Random numbers from the beta prime distribution.
}
\description{
RNG for the beta prime distribution
}
\examples{
hist(rbetaprime(100, mu = 1, phi = 2))
}
