% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{rgompertz}
\alias{rgompertz}
\title{RNG function for the Gompertz distribution, with Median parametrization.}
\usage{
rgompertz(n, mu, beta)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median parameter}

\item{beta}{Scale parameter}
}
\value{
A Gompertz distributed RNG vector of size n
}
\description{
RNG function for the Gompertz distribution, with Median parametrization.
}
\examples{
hist(rgompertz(n = 100, mu = 2, beta = 0.1))
}
