% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitnormal.R
\name{rlogitnormal}
\alias{rlogitnormal}
\title{Logitnormal RNG-function in median parametrization.}
\usage{
rlogitnormal(n, mu = 0, sigma = 1)
}
\arguments{
\item{n}{number of observations}

\item{mu}{Median parameter, mu unbound, mu already logit transformed}

\item{sigma}{Sigma shape parameter, sigma > 0}
}
\value{
n Logitnormal distributed samples
}
\description{
Logitnormal RNG-function in median parametrization.
}
\examples{
hist(rlogitnormal(100, 0.5, 2))
}
