% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lomax.R
\name{rlomax}
\alias{rlomax}
\title{RNG function for the Lomax distribution, with Mean parametrization.}
\usage{
rlomax(n, mu, alpha)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Median argument of Lomax}

\item{alpha}{Eta argument of Lomax}
}
\value{
A Lomax distributed RNG vector of size n
}
\description{
RNG function for the Lomax distribution, with Mean parametrization.
}
\examples{
y <- bayesim::rlomax(n, mu = 2, eta = 2)
hist(log(y), main = c(paste("Median:", mean(y)), " for RNG of high starting Lomax(mu=2, eta=0.1)"))
}
