% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{rng_lookup}
\alias{rng_lookup}
\title{Lookup function for RNGs via string identifier}
\usage{
rng_lookup(family)
}
\arguments{
\item{family}{String identifier of the likelihood family to get an RNG for.}
}
\value{
The RNG function.
}
\description{
Lookup function for RNGs via string identifier
}
\examples{
rng_lookup("gamma")
do.call(rng_lookup("gaussian"), list(n = 100, mean = 0, sd = 1))
}
