% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitlindley.R
\name{runit_lindley}
\alias{runit_lindley}
\title{Unit-Lindley RNG function}
\usage{
runit_lindley(n, mu)
}
\arguments{
\item{n}{number of observations}

\item{mu}{Mean, mu e (0, 1)}
}
\value{
n samples drawn from the Unit Lindley distribution
}
\description{
Unit-Lindley RNG function
}
\examples{
hist(runit_lindley(100, 0.5))
}
