% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifted_inv_gaussian.R
\name{shifted_inv_gaussian}
\alias{shifted_inv_gaussian}
\title{Title}
\usage{
shifted_inv_gaussian(
  link = "1/mu^2",
  link_shape = "log",
  link_ndt = "identity"
)
}
\arguments{
\item{link_ndt}{}
}
\description{
Title
}
