% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex}
\alias{simplex}
\title{Simplex brms-implementation in median parametrization.}
\usage{
simplex(link = "logit", link_sigma = "identity")
}
\arguments{
\item{link}{Link function for function}

\item{link_sigma}{Link function for sigma argument}
}
\value{
brms Beta-Custom distribution family
}
\description{
Simplex brms-implementation in median parametrization.
}
\examples{
a <- rnorm(1000)
data <- list(a = a, y = rsimplex(1000, brms::inv_logit_scaled(0.5 * a + 1), 2))
fit <- brms::brm(
  formula = y ~ 1 + a, data = data,
  family = simplex(), stanvars = simplex()$stanvars,
  refresh = 0
)
plot(fit)
}
