% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softplusnormal.R
\name{softplusnormal}
\alias{softplusnormal}
\title{Custom BRMS family Softplus in median parametrization.}
\usage{
softplusnormal(link = "identity", link_sigma = "log")
}
\arguments{
\item{link}{Link function argument (as string) for Median argument. Left as identity!}

\item{link_sigma}{Link function argument (as string) for Shape argument}
}
\value{
Softplus BRMS model-object
}
\description{
Custom BRMS family Softplus in median parametrization.
}
\examples{
# Running the example might take a while and may make RStudio unresponsive.
# Just relax and grab a cup of coffe or tea in the meantime.
a <- rnorm(1000)
data <- list(a = a, y = rsoftplusnormal(1000, 0.5 * a + 1, 2))
# BBmisc::surpressAll necassary to keep test output clean
BBmisc::suppressAll({
  fit1 <- brms::brm(y ~ 1 + a,
    data = data, family = softplusnormal(),
    stanvars = softplusnormal()$stanvars
  )
})
plot(fit1)
}
