% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helper.R
\name{test_rng_asym}
\alias{test_rng_asym}
\title{Test if an RNG asymptotically approaches the true location}
\usage{
test_rng_asym(
  rng_fun,
  metric_mu,
  n_samples = c(10, 10000),
  mu_list,
  aux_list,
  mu_link = identity,
  allowed_failures = 0.05
)
}
\arguments{
\item{rng_fun}{RNG function under test}

\item{metric_mu}{Metric to be used on RNG data (usually mean or median)}

\item{n_samples}{Default=c(10, 10000), sample sizes for rng test, len >= 2. Gets sorted in routine.}

\item{mu_list}{Metric data used as RNG argument and to be compared to (usually mean or median)}

\item{aux_list}{Auxiliary parameter value list.}

\item{mu_link}{Default=identity, optional link-function argument, for example}

\item{allowed_failures}{Default=0.05 marks, that 5 percent of all test cases are allowed
to fail for the whole test to still succeed.}
}
\value{
Success or failure with message

#@examples eps <- 0.001
mu_list <- seq(from = 1 + eps, to = 20, length.out = 10)
phis <- seq(from = 2 + eps, to = 20, length.out = 10)
# if working as expected, this test should not print any errors
result <- bayesfam:::test_rng_asym(
  rng_fun = rbetaprime,
  metric_mu = mean,
  mu_list = mu_list,
  aux_list = phis,
)
print(result)
}
\description{
Is currently not in a reliable state useful for testing. Needs more thought.
}
